+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## WIP
+
+### Features
+
+- add Prometheus metrics endpoint to the receiver app
+
 ## 0.1.8 - 2026-02-05
 
 ### Features